Added reverse-complement Haskell and Clojure program, and some other changes
Found that in Clojure, using a literal map as a function, repeatedly,
causes that map to be recreated every time that code is executed.
Much faster to assign that map to a Var and/or bind it to a name with
let.  That improves my earlier Clojure reverse-complement programs
significantly, and improves the earlier Clojure k-nucleotide entries
somewhat.

Figured out how to get Clojure k-nucleotide and reverse-complement
entries to read from stdin instead of a file.  Fixed up the shell
scripts invoking them to match.

Added clojure-contrib jar file to the classpath used when running
Clojure programs.

Copied Haskell entry for reverse-complement from language shootout web
site.