I finally worked around some limitations where bit-shift-left with a first arg bigger than a 32-bit int and a last arg that was a variable caused reflection warnings, and thus very slow processing in an inner loop.  I changed it so that the shift amount was always a compile-time constant, which somehow causes it to avoid the reflection, and go much faster.
This version also has some support for parallel threads, but for some
reason I have not yet tracked down it seems to use more total CPU time
when run on a 2-core machine with parallelism of 3, than if I run it
with a parallelism of 1.  The parallel threads seem to be interfering
with each other somehow.