Modified --jvm-gc-stats and added --jruby-gc-stats to match the form and behavior of measureproc.
Changes to how GC summary statistics are displayed when --verbose is
given.  Added display of the number of GCs that occurred.  Changed
definition of max live to be closer to the meaning one might expect,
i.e. the total allocated space after a GC is complete.  The total
allocated space before a GC begins is now called 'max allocated'.  Max
live and max allocated are not printed if there were 0 GC invocations,
since they really don't measure anything in that case.